Add workspace id, Update edit_task,
Add workspace id in the url link so that we can differentiate workspaces as different workspaces can edit their own tasks and have their own fixed template tasks.

Update function updateTemplateTask because Template tasks data should not be editable because they are used as a fixed value per user and workspaces, so I added a new DB table where we can get the edited tasks and put it in the new table and keep the original template tasks data.

The query overwrites any old data if the new data that will be inserted is the same as the old data's task_id and workspace_id.